% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\title{Deprecated}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The column \code{company_id} is now deprecated in all \code{toy_*companies()}. The
latest data uses \code{companies_id} and is available to you automatically when
you update the tiltIndicator package.
}
\keyword{internal}
