% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_files.R
\name{deprecated_path}
\alias{deprecated_path}
\title{Paths to deprecated datasets}
\usage{
deprecated_path(file)
}
\arguments{
\item{file}{File.}
}
\value{
A character.
}
\description{
This package comes bundled with a number of toy datasets in its
'inst/extdata' directory. Use \code{toy_files()} to list all the available
datasets and \code{toy_path()} to retrieve the path to one example.
}
\examples{
list.files(deprecated_path(""))
}
\keyword{internal}
