% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_path.R
\name{toy_datasets}
\alias{toy_datasets}
\alias{toy_dataset}
\title{Get path to toy datasets}
\usage{
toy_datasets(pattern = NULL)

toy_dataset(file)
}
\arguments{
\item{pattern}{A regular expression of filenames to match. If \code{NULL} all
available files are returned. listed.}

\item{file}{File.}
}
\value{
A character.
}
\description{
This package comes bundled with a number of toy datasets in its
'inst/extdata' directory. Use toy_datasets() to list all the available
datasets and toy_dataset() to retrieve the path to one example.
}
\examples{
files <- toy_datasets()
files

toy_dataset(files[[1]])
}
