% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{toy_emissions_profile_products}
\alias{toy_emissions_profile_products}
\alias{toy_emissions_profile_upstream_products}
\title{Deprecated \verb{toy_emissions_profile*()}}
\usage{
toy_emissions_profile_products()

toy_emissions_profile_upstream_products()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
toy_emissions_profile_products()
# ->
toy_emissions_profile_products_ecoinvent()

toy_emissions_profile_upstream_products()
# ->
toy_emissions_profile_upstream_products_ecoinvent()
}
\keyword{internal}
