% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_files.R
\name{toy_files}
\alias{toy_files}
\alias{toy_path}
\title{Get path to toy datasets}
\usage{
toy_files(pattern = NULL)

toy_path(file)
}
\arguments{
\item{pattern}{A regular expression of filenames to match. If \code{NULL} all
available files are returned. listed.}

\item{file}{File.}
}
\value{
A character.
}
\description{
This package comes bundled with a number of toy datasets in its
'inst/extdata' directory. Use \code{toy_files()} to list all the available
datasets and \code{toy_path()} to retrieve the path to one example.
}
\examples{
toy_files()

toy_path(toy_files()[[1]])

toy_emissions_profile_products()
}
