% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltToyData-package.R
\docType{package}
\name{tiltToyData-package}
\alias{tiltToyData}
\alias{tiltToyData-package}
\title{tiltToyData: Toy Datasets For 'TILT'}
\description{
Toy datasets for 'TILT'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/tiltToyData/}
  \item \url{https://github.com/2DegreesInvesting/tiltToyData}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltToyData/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
