% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_datasets.R
\name{toy_datasets}
\alias{toy_datasets}
\alias{toy_emissions_profile_any_companies}
\alias{toy_emissions_profile_products}
\alias{toy_emissions_profile_upstream_products}
\alias{toy_sector_profile_any_scenarios}
\alias{toy_sector_profile_companies}
\alias{toy_sector_profile_upstream_companies}
\alias{toy_sector_profile_upstream_products}
\title{Toy datasets}
\usage{
toy_emissions_profile_any_companies()

toy_emissions_profile_products()

toy_emissions_profile_upstream_products()

toy_sector_profile_any_scenarios()

toy_sector_profile_companies()

toy_sector_profile_upstream_companies()

toy_sector_profile_upstream_products()
}
\description{
Toy datasets
}
