% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_datasets.R, R/toy_files.R
\name{toy_emissions_profile_any_companies}
\alias{toy_emissions_profile_any_companies}
\alias{toy_emissions_profile_products_ecoinvent}
\alias{toy_emissions_profile_upstream_products_ecoinvent}
\alias{toy_sector_profile_any_scenarios}
\alias{toy_sector_profile_companies}
\alias{toy_sector_profile_upstream_companies}
\alias{toy_sector_profile_upstream_products}
\alias{toy_files}
\alias{toy_path}
\title{Paths to toy datasets}
\usage{
toy_emissions_profile_any_companies()

toy_emissions_profile_products_ecoinvent()

toy_emissions_profile_upstream_products_ecoinvent()

toy_sector_profile_any_scenarios()

toy_sector_profile_companies()

toy_sector_profile_upstream_companies()

toy_sector_profile_upstream_products()

toy_files(pattern = NULL)

toy_path(file)
}
\arguments{
\item{pattern}{A regular expression of filenames to match. If \code{NULL} all
available files are returned. listed.}

\item{file}{File.}
}
\value{
A character.
}
\description{
This package comes bundled with a number of toy datasets in its
'inst/extdata' directory. Use \code{toy_files()} to list all the available
datasets and \code{toy_path()} to retrieve the path to one example.
}
\examples{
toy_emissions_profile_any_companies()

toy_emissions_profile_products_ecoinvent()

toy_emissions_profile_upstream_products_ecoinvent()

toy_sector_profile_any_scenarios()

toy_sector_profile_companies()

toy_sector_profile_upstream_companies()

toy_sector_profile_upstream_products()
toy_files()

toy_path(toy_files()[[1]])

}
